[jOOQ/jOOQ#12494] Rename Record.changed() to Record.touched()

+++ jOOQ-meta/src/main/java/org/jooq/meta/duckdb/system/information_schema/tables/KeyColumnUsage.java
@@ -92,7 +92,7 @@ public class KeyColumnUsage extends TableImpl<Record> {
      * The column
      * <code>system.information_schema.key_column_usage.ordinal_position</code>.
      */
-    public final TableField<Record, Integer> ORDINAL_POSITION = createField(DSL.name("ordinal_position"), SQLDataType.INTEGER, this, "");
+    public final TableField<Record, Long> ORDINAL_POSITION = createField(DSL.name("ordinal_position"), SQLDataType.BIGINT, this, "");
 
     /**
      * The column
